Refactor ListItemLink to consolidate collapsible functionality and remove ListItemLinkCollapsible component
