+++ src/caretogether-pwa/src/Shell/ListItemLink.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import ListItem from '@mui/material/ListItem';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
 import { useMatch } from 'react-router';
@@ -8,56 +7,97 @@ import {
   LinkProps as RouterLinkProps,
 } from 'react-router-dom';
 import { DistributiveOmit } from '@mui/types';
-import { OpenInNew } from '@mui/icons-material';
+import { Collapse, List, ListItemButton } from '@mui/material';
+import { ExpandLess, ExpandMore } from '@mui/icons-material';
 
-// Reference: https://material-ui.com/guides/composition/
-
-interface ListItemLinkProps {
+interface ListItemLinkCollapsibleProps {
   icon?: React.ReactElement;
   primary: string;
   to: string;
-  newTab?: boolean;
-  darkColor?: boolean;
+  subitems?: {
+    label: string;
+    isActive: boolean;
+    onClick: () => void;
+  }[];
+  defaultOpen?: boolean;
   className?: string;
 }
 
-function ListItemLink(props: ListItemLinkProps) {
-  const { icon, primary, to, newTab, darkColor } = props;
+export function ListItemLink(props: ListItemLinkCollapsibleProps) {
+  const { icon, primary, to, subitems, defaultOpen } = props;
 
-  const match = useMatch({
-    path: to,
-  });
+  const selected =
+    useMatch({
+      path: to,
+    }) !== null;
 
   const renderLink = React.useMemo(
     () =>
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       React.forwardRef<any, DistributiveOmit<RouterLinkProps, 'to'>>(
-        (itemProps, ref) => (
-          <RouterLink
-            to={to}
-            ref={ref}
-            {...itemProps}
-            target={newTab ? '_blank' : undefined}
-          />
-        )
+        (itemProps, ref) => <RouterLink to={to} ref={ref} {...itemProps} />
       ),
-    [newTab, to]
+    [to]
   );
+
+  const [collapsibleOpen, setCollapsibleOpen] = React.useState(
+    defaultOpen || false
+  );
+
+  const collapseIcon = collapsibleOpen ? <ExpandLess /> : <ExpandMore />;
+
+  const hasSubitems = subitems && subitems.length > 0;
 
   return (
-    <li className={props.className}>
-      <ListItem
-        button
-        component={renderLink}
-        selected={match !== null}
-        sx={{ paddingLeft: 1.5, color: darkColor ? '#555' : '#fff8' }}
+    <>
+      <li className={props.className}>
+        <ListItemButton
+          component={renderLink}
+          onClick={() => selected && setCollapsibleOpen(!collapsibleOpen)}
+          selected={selected}
+          sx={{ paddingLeft: 1.5 }}
+        >
+          {icon ? (
+            <ListItemIcon sx={{ color: selected ? '#fff' : '#fff8' }}>
+              {icon}
+            </ListItemIcon>
+          ) : null}
+
+          <ListItemText
+            primary={primary}
+            sx={{ marginLeft: -2, color: selected ? '#fff' : '#fff8' }}
+          />
+
+          {hasSubitems && collapseIcon}
+        </ListItemButton>
+      </li>
+
+      <Collapse
+        in={hasSubitems && collapsibleOpen}
+        timeout="auto"
+        unmountOnExit
       >
-        {icon ? <ListItemIcon>{icon}</ListItemIcon> : null}
-        <ListItemText primary={primary} sx={{ marginLeft: -2 }} />
-        {newTab && <OpenInNew fontSize="small" sx={{ marginLeft: 1 }} />}
-      </ListItem>
-    </li>
+        <List component="div" disablePadding>
+          {subitems?.map((item) => (
+            <li className="ph-unmask" key={item.label}>
+              <ListItemButton
+                selected={item.isActive}
+                sx={{ paddingLeft: 1.5, color: '#fff8' }}
+                onClick={item.onClick}
+              >
+                {/* {icon ? <ListItemIcon>{icon}</ListItemIcon> : null} */}
+                <ListItemText
+                  primary={item.label}
+                  sx={{
+                    marginLeft: 6,
+                    color: item.isActive ? '#fff' : undefined,
+                  }}
+                />
+              </ListItemButton>
+            </li>
+          ))}
+        </List>
+      </Collapse>
+    </>
   );
 }
-
-export { ListItemLink };
